backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq>=0.11.0
python-dotenv>=1.0.0
pypdf==4.3.1
python-multipart==0.0.9
pydantic==2.8.2
sentence-transformers==3.3.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-text-splitters==0.3.3
faiss-cpu>=1.7.4
langgraph>=0.2.0
python-docx==1.1.2
numpy
